Add main function and check if the fist argument is a path
If the parameter -h is passed as the first argumen, then the current
working directory is used as the path. Otherwise, the program checks
that the provided string correspond to an existing path on the OS.

Finally, the last / character is stripped from the path if it was
provided in the argument.